refactor: modularize UI components
